requirements.txt
langchain
langchain-community
langchain-groq
langgraph
langchain-text-splitters
ddgs
streamlit
python-dotenv